LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/worker-interception-redirect.https-expected.txt



Harness Error (TIMEOUT), message = null

PASS initialize global state
FAIL request to sw1 scope gets network redirect to sw2 scope assert_equals: expected "sw1 saw importScripts from the worker" but got "importScripts: served from network"
FAIL request to sw1 scope gets network redirect to out-of-scope assert_equals: expected "sw1 saw importScripts from the worker" but got "importScripts: served from network"
TIMEOUT request to sw1 scope gets service-worker redirect to sw2 scope Test timed out
NOTRUN request to sw1 scope gets service-worker redirect to out-of-scope
NOTRUN cleanup global state
